Refactor test_integrate.py.
- Move CustomGaussOld from basic.py to test_integrate.py.
- Import BasePDF class directly instead of basepdf module.
- Delete unused variables.
- Replace usage of math.inf with np.inf.
- Replace usage of z.pi with znp.pi.
- Remove now unused math import from basic.py and test_integrate.py.
